requirements.txt
fastapi
uvicorn
pyjwt
passlib
